Threading in python sucks
Yes, threading in general is all about pain, but in python, you're up
for interrupts being directed to random threads, weird deadlocks, that
from a point of logic can't occur and what not.

I tried to ease the pain by using pykka, a library for actors in python,
which handles some of the problems.

Please try to

* Use asynchronous requests wherever possible - async can not be a
  deadlock.
* not use threads. Actors are the new threads. An actor can receive
  messages and send them to actors they know. It's like OOP on
  hyper-drive.